26-OCI-function/my-func/requirements.txt
fdk
pip
mysql-connector-python
netifaces 
oci>=2.2.18
